requirements.txt
aiohttp==3.9.3
hydra-core==1.3.2
omegaconf==2.3.0
tenacity==8.2.3
pandas==2.2.1
anthropic==0.16.0
tiktoken==0.6.0
transformers==4.38.1
openai==1.13.3
gitpython==3.1.40